Refactor card storage initialization and copying default sets
This commit refactors the initialization process of the card storage service. The `init` method now checks if the home directory exists and creates it if it doesn't. Additionally, the `copyFolderSync` method is called to copy the default sets to the home directory.

Also, remove commented code in DownloadDefaultSetsDialog.vue

The commented code in the `onMounted` function of the `DownloadDefaultSetsDialog.vue` component has been removed. This code was previously used to handle the dialog visibility and download progress, but it is no longer needed.

+++ src/electron/services/card-storage-service.ts
@@ -42,8 +42,9 @@ export class CardsStorage extends ICloudStorage {
   init () {
     if (!existsSync(HOME_DIR)) {
       mkdirSync(HOME_DIR);
-      this.copyFolderSync(DEFAULT_SETS, HOME_DIR);
     }
+    this.copyFolderSync(DEFAULT_SETS, HOME_DIR);
+
   }
 
   async showItemInFolder (path: string) {
@@ -330,10 +331,13 @@ export class CardsStorage extends ICloudStorage {
     readdirSync(srcPath).forEach((file: string) => {
       const srcFilePath = join(srcPath, file);
       const destFilePath = join(destPath, file);
-      console.log(destFilePath);
+      
       if (lstatSync(srcFilePath).isDirectory()) {
         this.copyFolderSync(srcFilePath, destFilePath);
       } else {
+        if (existsSync(destFilePath)) {
+         return;
+        }
         copyFileSync(srcFilePath, destFilePath);
       }
     });
